backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mongoengine
pymysql
sqlalchemy
pymongo
cryptography
python-jose[cryptography]
passlib
bcrypt<4.0
redis
celery